creating a stripe card charge

+++ carts/models.py
@@ -6,7 +6,7 @@ from products.models import Item, Variation
 
 
 class Cart(models.Model):
-	total = models.DecimalField(max_digits=100, decimal_places=2, default=0.00)
+	total = models.FloatField(default=0.00)
 	created = models.DateTimeField(auto_now_add=True, auto_now=False)
 	updated = models.DateTimeField(auto_now_add=False, auto_now=True)
 	active = models.BooleanField(default=True)
@@ -27,7 +27,7 @@ class CartItem(models.Model):
 	item = models.ForeignKey(Item, on_delete=models.CASCADE)
 	quantity = models.IntegerField(default=1)
 	notes = models.TextField(null=True, blank=True)
-	line_total = models.DecimalField(default=0.00,max_digits=1000, decimal_places=2)
+	line_total = models.FloatField(default=0.00)
 	created = models.DateTimeField(auto_now_add=True, auto_now=False)
 	updated = models.DateTimeField(auto_now_add=False, auto_now=True)
